moved helper functions to 'Shared' folder, updated package.json
