+++ client/src/components/R&R/ReviewsSort.jsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import $ from 'jquery';
 import axios from 'axios';
 import styled from 'styled-components';
-import handler from '../../helpers/reviewhandler.js';
+import handler from '../Shared/reviewhandler.js';
 
 class ReviewsSort extends React.Component {
   constructor(props) {
